fixed #1, int type cannot handle calculated total number of iteration when it goes very big

+++ trunk/AbstractString.cpp
@@ -25,7 +25,7 @@ AbstractString::~AbstractString()
 }
 
 
-int AbstractString::total()
+double AbstractString::total()
 {
     if (computedTotal == -1)
         computedTotal = computeTotal();
